Progress made so far to simulate SARS-CoV-2 amplicons with PCR artefacts
Pandemic-scale SARS-CoV-2 SNPs under positive selection are simulated and sampled with VGsim. PhastSim is then used to introduce neutral mutations along the tree. A custom script extracts amplicon sequences in each simulated genome and samples sequencing coverages conditional on mismatches in the primer-binding regions. ART is then used to generate reads from the amplicons, the reads are combined for each simulated genome and reads passed to viridian for assembly.