Changed the direction in which mismatches are induced in overlapping amplicons and summary amplicon statistics are outputted
Previously I was inducing mismatches directly in primer binding sites. This has now been changed so that mismatches can be induced in the primer binding site of the upstream amplicon of the tiling scheme, replacing the primer binding site of the current amplicon with the primer sequence itself. Summary statistics about which error mode has been applied to each amplicon is now collected too.