Significant rewrite of the error mode simulation code and added masking of low coverage amplicons in assemblies
Previously I had been inducing SNPs in the primer binding regions of amplicons in order to simulate the error modes and missing out SNPs in the 3' primer site.  These changes remove the option to retrospectively simulate SNPs and instead checks for SNPs simulated by the previous steps in the pipeline without supplementing them. The code is now reduced, more modular and efficient. I also now replace low coverage regions of sequence with Ns to get a more accurate assessment of the performance of viridian.